Major Change form RGBW to HSVA Color Space

+++ include/Uhrtypes/DE10x11.alternative.frame.hpp
@@ -24,7 +24,7 @@ public:
 
     //------------------------------------------------------------------------------
 
-    virtual inline const uint16_t numPixelsFrameMatrix() { return 48; }
+    virtual inline const uint8_t numPixelsFrameMatrix() { return 48; }
 
     //------------------------------------------------------------------------------
